requirements.txt
langchain>=0.1.0
langchain-core>=0.1.10
langchain-community>=0.0.10
langchain-text-splitters>=1.0.0
langchain-google-genai>=2.0.5
langchain-huggingface>=0.0.1
sentence-transformers>=2.3.0
pinecone-client>=3.0.0
langchain-pinecone>=0.1.0
networkx>=3.2.0
pyvis>=0.3.2
python-dotenv>=1.0.0
duckduckgo-search>=4.0.0
beautifulsoup4>=4.12.3
requests>=2.31.0
lxml>=5.1.0
aiohttp>=3.9.0
pandas>=2.1.4
numpy>=1.26.3
langchain-experimental>=0.0.47
pytest>=7.4.3
pytest-asyncio>=0.23.3
pydantic>=2.5.3
httpx>=0.26.0
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
trafilatura>=1.6.0
spacy>=3.7.0
hf_xet
deepagents>=0.1.0
pypdf>=3.17.0
python-docx>=1.1.0
python-pptx>=0.6.23
openpyxl>=3.1.2

